# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Geography Quiz platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes/{id}/questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions for a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a question for a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Quiz not found
  /quizzes/{id}/questions/import-by-category:
    post:
      operationId: importQuestionsByCategory
      tags: [questions]
      summary: Copy all questions tagged with a category (or any descendant) into this quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportQuestionsByCategoryBody"
      responses:
        "200":
          description: Import summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportQuestionsByCategoryResult"
        "404":
          description: Quiz or category not found
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a single question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Question:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        explanation:
          type: string
        funFact:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        orderIndex:
          type: integer
        categories:
          type: array
          items:
            $ref: "#/components/schemas/QuestionCategory"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - quizId
        - text
        - options
        - correctOption
        - explanation
        - funFact
        - imageUrl
        - orderIndex
        - categories
        - createdAt
        - updatedAt
    QuestionCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
      required:
        - id
        - name
        - slug
    CreateQuestionBody:
      type: object
      properties:
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        explanation:
          type: string
        funFact:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        orderIndex:
          type: integer
        categoryIds:
          type: array
          items:
            type: integer
      required:
        - text
        - options
        - correctOption
        - explanation
        - orderIndex
    UpdateQuestionBody:
      type: object
      properties:
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        explanation:
          type: string
        funFact:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        orderIndex:
          type: integer
        categoryIds:
          type: array
          items:
            type: integer
    ImportQuestionsByCategoryBody:
      type: object
      properties:
        categoryId:
          type: integer
      required:
        - categoryId
    ImportQuestionsByCategoryResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        categoryName:
          type: string
      required:
        - imported
        - skipped
        - categoryName
...